feat(proposals): initialize default topic filter state (#6744)
# Motivation

We want to track the topics filter for SNS proposals in the same way we
track other filters, such as types and status. The store was already
updated in #6732. This PR initializes the filter similarly to the
`topics` filter, providing a default value when the store lacks any
value. A follow-up PR will load the store from local storage, ensuring a
better user experience.

[NNS1-3666](https://dfinity.atlassian.net/browse/NNS1-3666)

# Changes

- Sets the default value for the topics field in the sns
`snsFiltersStore` when no filter is defined.

# Tests

- Adds a test for the existing code regarding the default value of the
`type` filter.
- Adds tests for the new `topics` filter.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

Next PR: #6745

[NNS1-3666]:
https://dfinity.atlassian.net/browse/NNS1-3666?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ