feat(proposals): add sorted topic filter list util (#6745)
# Motivation

We want to track the topics filter for SNS proposals in the same way we
track other filters, such as types and status.

This PR creates a new utility to generate a list of filters based on
existing topics and persisted filters. The filters are sorted according
to the following logic:

-  First, critical topics are sorted alphabetically.
-  Second, non-critical topics are sorted alphabetically.
-  Lastly, a special filter shows all proposals without a topic.

[NNS1-3666](https://dfinity.atlassian.net/browse/NNS1-3666)

# Changes

- New utility to create a list of filters from topics and saved filters.
If there are no topics, the filter list will be empty.

# Tests

- Unit tests for this new util.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

Prev. PR: #6744 | Next PR: #6746

[NNS1-3666]:
https://dfinity.atlassian.net/browse/NNS1-3666?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ