feat(proposals): util for topic-based list filtering request payload  (#6762)
# Motivation

We want users to view proposals filtered by topic. To accomplish this,
we need to construct the appropriate payload when calling
`list_proposals` in the governance canister.

Note: When the user selects the special entry in the filter for
proposals without a topic, we must pass a null value to the API. More
info
[here](https://github.com/dfinity/ic/blob/master/rs/sns/governance/src/gen/ic_sns_governance.pb.v1.rs#L3200)

[NNS1-3666](https://dfinity.atlassian.net/browse/NNS1-3666)

# Changes

- New util to compose the payload required by `list_proposals` to filter
by topic.

# Tests

- Unit tests for the new util.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3666]:
https://dfinity.atlassian.net/browse/NNS1-3666?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ