feat(store): add minimum dissolve delay from network economics (#6781)
# Motivation

We want to start using the minimum dissolve delay from the API instead
of the hardcoded value `NNS_MINIMUM_DISSOLVE_DELAY_TO_VOTE`. This change
is necessary because the value may change in the near future, and we
want the application to respond accordingly.

This first PR introduces a store that retrieves this value from the
network economics API.

It defaults to the existing value while the API defines it:
[thread](https://dfinity.slack.com/archives/C039M7YS6F6/p1746198021862419?thread_ts=1746197917.500179&cid=C039M7YS6F6).

[NNS1-3762](https://dfinity.atlassian.net/browse/NNS1-3762)

# Changes

- Derived store to expose the `neuronMinimumDissolveDelayToVoteSeconds`
value.

# Tests

- Adds unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.

[NNS1-3762]:
https://dfinity.atlassian.net/browse/NNS1-3762?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ