refactor(topics): use derived store for topic list access (#6787)
# Motivation

We want to apply the same sorting used for the proposals filter #6745 in
other parts of the application. This requires that consumers of the list
of topics retrieve them from the derived store
`createSnsTopicsProjectStore` instead of directly from the topics store.

This PR prepares the component to consume this store, so once the store
returns the sorted topics, they will be automatically used.
Additionally, it has been migrated to Svelte 5.

[NNS1-3792](https://dfinity.atlassian.net/browse/NNS1-3792)

# Changes

- Replace direct access to `snsTopicsStore` with
`createSnsTopicsProjectStore`.
- Refactor the component to use runes instead of the reactive syntax
`$:`.

# Tests

- Tests should pass as before since there were no functional changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3792]:
https://dfinity.atlassian.net/browse/NNS1-3792?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ