fix(ui): add followee button visibility after addition (#6767)
# Motivation

We have an issue with the "Add Followee" button, which becomes partially
hidden after adding a followee (#6715).

This problem relates to how Svelte calculates dynamic spaces for our
`Collapsible` component. As this flow will be change in a near future to
to use topic following similar to how the SNS governance is handled, the
propose solution is to force `Collapsible` to re-calculate its height by
updating one prop.

Before:


https://github.com/user-attachments/assets/e7a5d55c-56d6-41de-981c-e9a54d0c67ce

After:


https://github.com/user-attachments/assets/1d12ed49-3f3a-4494-ae45-925ab5f59d69

Closes #6715.

# Changes

- Temporarily change a prop to retrigger the height calculation.
- Refactor to Svelte 5.

# Tests

- Update unit tests
- Tested in
[devenv](https://qsgjb-riaaa-aaaaa-aaaga-cai.yhabib-ingress.devenv.dfinity.network)

# Todos

- [ ] Add entry to changelog (if necessary).