feat(delegation-state-column): New NeuronVoteDelegationCell component (#6803)
# Motivation

To improve the overview of a user’s vote delegation state, it’s planned
to display the delegation status icon directly in the neuron table by
adding a new column: “Vote Delegation” (see screenshot).

This PR adds a new component that will be used as the cell content for
this column in both the NNS and SNS tables.

[Draft MVP PR](https://github.com/dfinity/nns-dapp/pull/6802/files)

[Demo](https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/)

![Screenshot 2025-05-08 at 13 38
40](https://github.com/user-attachments/assets/700ce2e2-2581-4d59-90a8-97dcdce72357)

# Changes

- Add voteDelegationState to TableNeuron type.
- New `NeuronVoteDelegationCell` component.

# Tests

- Only PO is created. The spec will be added later.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.