feat(portfolio): implement pre-swap SNS proposal card (#6811)
# Motivation

We want to display proposals for new SNS projects on the Portfolio page
that were approved before the swap begins. This first PR introduces a
new type of card for these cases.

Note: There are some deviations from the design: there is no project
link, as it is not always available, and we will reuse the same
remaining time format as the other cards.

Design:

![Screenshot 2025-05-09 at 08 25
42](https://github.com/user-attachments/assets/baae3600-868e-4f6c-ba1b-2d98b471e665)

Implementation:

![Screenshot 2025-05-09 at 08 26
44](https://github.com/user-attachments/assets/dc43789e-e1ee-42b7-8ebc-7e5d3f66f1af)

[NNS1-3639](https://dfinity.atlassian.net/browse/NNS1-3639)

# Changes

- New `AdoptedProposalCard` component.

# Tests

- Unit tests for the new component.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3639]:
https://dfinity.atlassian.net/browse/NNS1-3639?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ