feat(api): add new method listSubaccounts (#6815)
# Motivation

We want to utilize the new function exposed by ic-js to [list
subaccounts](https://github.com/dfinity/ic-js/pull/906), enabling users
to easily retrieve the list of all subaccounts for an ICRC canister.

Note: This PR introduces a new canister object based on the
NextGeneration Index canister. The existing one will be deprecated
eventually, so a follow-up PR will migrate the existing methods to the
new canister.

[NNS1-3793](https://dfinity.atlassian.net/browse/NNS1-3793)

# Changes

- Add a new canister object for the Next Generation IndexCanister.  
- Add an API function to call `list_subaccounts`.

# Tests

- Add unit tests for the API function.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3793]:
https://dfinity.atlassian.net/browse/NNS1-3793?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ