feat(portfolio): integrate adopted SNS proposal cards (#6822)
# Motivation

We want to display proposals for new SNS projects on the Portfolio page
that were approved before the swap begins. This PR utilizes the new card
introduced in #6811 to showcase adopted proposals for new projects.

The order for displaying cards within the stacked card component is as
follows:
1. Ongoing swaps
2. Proposals for new projects
3. Adopted proposals for new projects.

[NNS1-3639](https://dfinity.atlassian.net/browse/NNS1-3639)

# Changes

- Update Portfolio page to expected a new prop for adopted proposals.
- Sort adopted proposals based on the swap start.

# Tests

- Unit tests for the new prop in the page.
- Unit tests for checking the sorting is applied to each group of cards.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.

[NNS1-3639]:
https://dfinity.atlassian.net/browse/NNS1-3639?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ