feat(utils): add hidden balance character generator (#6843)
# Motivation

We want to introduce a toggle that allows users to hide or show their
balances on the main pages. When hidden, the design requires the
"hidden" character to be displayed multiple times, depending on the
context and available space. This PR adds a new utility to generate
these "hidden" characters.

[NNS1-3721](https://dfinity.atlassian.net/browse/NNS1-3721)

# Changes

-  Added a new utility.

# Tests

-  Added unit tests.

# Todos

-  [ ] Add an entry to the changelog (if necessary)
Not necessary

[NNS1-3721]:
https://dfinity.atlassian.net/browse/NNS1-3721?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ