feat(ui): add privacy-aware amount display (#6860)
# Motivation

We want to introduce a toggle that allows users to hide or show their
balances on the main pages. This PR adds a new component that determines
whether to display the provided value or hide it with placeholder
characters. This component will later be used by various other
components.

# Changes

- New component `PrivacyAwareAmount`

# Tests

- Unit tests for the new component.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.