refactor(lists): migrate consumers of infinite scroll components to runes (#6859)
# Motivation

We want to upgrade the version of Gix in the nns-dapp to address an
issue with Infinite Scrolling. The latest version of GIX introduced a
breaking change to `InfiniteScroll` when migrating it to Svelte 5.

This PR prepares consumers of these components for the upgrade by
migrating them to runes.

# Changes

- Refactor `UiTransactionsList` to support runes.
- Refactor `NnsProposalsList` to support runes.
- Refactor `SnsProposalsList` to support runes.
- Refactor `UniverseWithActionableProposals` to support runes.

# Tests

- Pipeline should work as before.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.