feat(amounts): implement balance privacy display mode (#6852)
# Motivation

We want to introduce a toggle that allows users to hide or show their
balances on the main pages. This PR updates the `AmountWithUsd`
component to recognize the value of the hide balance store and render
the value either in a formatted manner or with a hidden character.


https://github.com/user-attachments/assets/4e5128fe-1311-498b-8492-c9ea06661d7b


# Changes

- Update `AmountWithUsd` to display the value either formatted or with
hidden characters based on `isBalancePrivacyOptionStore`.
- Extend `AmountDisplay` to include a prop that shows the value
formatted or with hidden characters.

# Tests

- Add unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.