feat(disburse-maturity): isEnoughMaturityToDisburse (#6897)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR introduces a function to calculate if the provided maturity is
enough for disbursement.

[Jira](https://dfinity.atlassian.net/browse/NNS1-3740)

# Changes

- Add `isEnoughMaturityToDisburse ` util.

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet