feat(ui): add inline copy feedback component (#6901)
# Motivation

We want to give in-place feedback to the users whenever they click any
element that gets copied into their clipboard. The nns-dapp currently
uses a component defined in Gix, but designers want to move away from
this version.

This first PR introduces the new component to the nns-dapp repository.


https://github.com/user-attachments/assets/e88ab47d-440d-4eeb-8c20-c379162ff663

[NNS1-3835](https://dfinity.atlassian.net/browse/NNS1-3835)

# Changes

- New component `Copy` with in-place feedback on successful copy.

# Tests

- Unit tests for the new component.
- Manually tested with other changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.

[NNS1-3835]:
https://dfinity.atlassian.net/browse/NNS1-3835?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ