feat(disburse-maturity): Disburse maturity modal (#6911)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR adds the "Disburse" modal to the neuron page.

**Out of scope**
- disbursing to sub-account
- disbursing to external addresses.

[Jira](https://dfinity.atlassian.net/browse/NNS1-3740)

# Changes

- New `NnsDisburseMaturityModal` component that heavily reuses
DisburseMaturityModal component.
- Display disburse maturity modal on page.

# Tests

- Added.
- Verified manually that a user can disburse maturity to the main
account.


https://github.com/user-attachments/assets/2550e073-0330-45db-92b3-70056e5b4131

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet