feat(disburse-maturity): Display total disbursements in progress (#6915)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR adds the field to the neuron page with the total disbursements
and the button to display more details.

![Screenshot 2025-06-02 at 18 43
26](https://github.com/user-attachments/assets/fef89302-67c6-456a-bffd-8c922538ab9f)

[Jira](https://dfinity.atlassian.net/browse/NNS1-3740)

# Changes

- New ViewActiveDisbursementsItemAction component.
- Display it on the neuron details page.

# Tests

- Added.
- Verified manually that a user can disburse maturity to the main
account.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet