chore(cleanup): prepare codebase for node 22 migration (#6940)
# Motivation

#6922 will upgrade Node from version 18 to 22. This will introduce
breaking changes, particularly in tests. To simplify the differences in
the main PR, changes that can be applied in isolation will be introduced
in separate PRs.

# Changes

- Add `.nvmrc` to `.gitignore` since the official version is defined in
`config.json`.
- Migrate `HideZeroNeuronsToggle` to runes.  
- Clean up `unsubscribeWatchMetrics`, as it was merely a placeholder
function.
- Migrate `CkBTCInfoCard.svelte` to runes and improve tests.  
- Replace the mock with a spy in `CkBTCTransactionModal.spec` to
maintain consistency with the rest of the test file.
- Remove spies on `queryNeurons` and `queryKnownNeurons` in
`NnsNeuronDetail.spec` since the API is fake, and in Node22, this will
pollute the rest of the tests.

# Tests

- Tests should pass as before.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?

Next PR: #6922