feat(disburse-maturity): destination for disbursement (#6941)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

The UI components already support custom destinations via account
identifier (hex string).
This PR ensures the correct API call is made using the provided
destination.

# Changes

- Added `toAccountIdentifier` field to the disburse maturity API.
- Passed the `toAccountIdentifier` value from the components to the API
layer for use in the request.

# Tests

- Updated.
- Tested locally that disbursement works with the custom account
identifier.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?