test(cleanup): remove conflicting mock verifications (#6943)
# Motivation

#6922 will upgrade Node from version 18 to 22. This will introduce
breaking changes, particularly in tests. To simplify the differences in
the main PR, changes that can be applied in isolation will be introduced
in separate PRs.

`blockAllCallsTo` is "installed" into a test file by providing a list of
modules against which we want to track unexpected calls. It verifies
that all calls made in the test cases are mocked. If any other calls are
made, it throws an error. This functionality no longer works the same
way in Node 22 because the order in which the `beforeEach` of
`blockAllCallsTo` and the test cases are executed has changed. Now, the
check can occur before the mock, resulting in an error being thrown.

# Changes

- Removes `blockAllCallsTo` from tests that have mock conflicts.

# Tests

- Tests should pass as before.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?

Prev. PR: #6940 | Next PR: #6922