Enable disburse maturity for NNS neurons (#6942)
# Motivation

To simplify the process of claiming matured rewards and avoid the extra
steps of creating and disbursing a new neuron, the disburse maturity
feature is introduced for NNS neurons (this approach is already used for
SNS neurons).

This PR switches the feature flag to enable disburse maturity for NNS
neurons.

# Changes

- Enable `ENABLE_DISBURSE_MATURITY` feature flag.
- Ran `./scripts/nns-dapp/test-config --update` to update config files.

# Tests

- Verified locally that the feature is enabled by default.
- Since the "Spawn Neuron" button was replaced with "Disburse":
  - Updated multiple screenshots due to the button change.
- Added an end-to-end test for disburse maturity to replace the spawn
neuron test (this also updates screenshots, as after disburse maturity
the user remains on the neuron details page).

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?