Dont hide empty neuron with disbursements (#6973)
# Motivation

After disbursing a neuron w/o maturity, it no longer appears in the
neurons table — even if there are ongoing maturity disbursements. This
is not desired, as it makes the disbursement status invisible to the
user, creating the impression that the maturity has simply disappeared.

To avoid unintended side effects, we decided to create a new derived
store, `nonEmptyNeuronStore` (which contains neurons with stake or
ongoing maturity disbursements), instead of modifying the
`definedNeuronStore`.

### How to reproduce:
**Preconditions:** unlocked Sns neuron with some maturity.
1. Click “Disburse” under maturity block and select 100%.
2. Click “Disburse” the neuron and disburse 100%.
3. Navigate to the NNS neuron table.

It’s expected that the neuron appears in the list.

[NNS1-3876](https://dfinity.atlassian.net/browse/NNS1-3876)

# Changes

- New `nonEmptyNeuronStore` derived store.
- Use it for the table neurons.

# Tests

- Test the new store.
- Test that neurons without stake, but with active maturity
disbursements are displayed in the table.
- Verified manually that the fix work.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3876]:
https://dfinity.atlassian.net/browse/NNS1-3876?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ