feat(analytics): add query parameter path conversion (#6974)
# Motivation

Plausible Analytics does not track query parameters. The nns-dapp uses
query parameters for nested pages such as `projects`, `universes`, and
`proposals`.

This PR introduces a utility that takes the path and query parameters
and composes a new URL with those query parameters as path elements.
This will replace the `enableAutoPageviews` feature to track the pages
that users visit.

Note: It doesn't convert all parameters, as some are user-specific, such
as `neuron`.

[NNS1-3874](https://dfinity.atlassian.net/browse/NNS1-3874)

# Changes

- Util to combine path and query params.

# Tests

- Unit tests for the new util.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3874]:
https://dfinity.atlassian.net/browse/NNS1-3874?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ