feat(analytics): implement custom page view tracking (#6982)
# Motivation

The `enableAutoPageviews` does not consider query parameters. Since the
dapps use them for nested navigation, such as projects or tokens, it is
necessary to manually trigger the events with custom values.

Additionally, the provider uses `location.href` for standard events if a
custom URL is not provided. The dapp must now provide it each time to
keep the pageEvent and the custom event in sync.

<img width="864" alt="Screenshot 2025-06-18 at 22 45 43"
src="https://github.com/user-attachments/assets/fe65e468-304f-4e05-bc0f-d21963939c15"
/>

[NNS1-3874](https://dfinity.atlassian.net/browse/NNS1-3874)

# Changes

- Replace automatic page view tracking with custom navigation hook
- Update `analytics.event`api to accept the url of the event.
- Updated events.

# Tests

- Updated tests for the new api.
- Manually tested

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3874]:
https://dfinity.atlassian.net/browse/NNS1-3874?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ