feat(store): add viewport size detection (#7062)
# Motivation

New designs for the nns-dapp define different DOM structures for mobile
and non-mobile desktops. Although CSS media queries can handle such
cases, they would create too many unnecessary DOM nodes. This PR
introduces a new store that exposes the viewport size and a derived
store to determine if it is mobile (less than 576px) or not.

Note: It hardcodes the value of the small breakpoint, but this will be
updated to a global constant later.

[NNS1-3922](https://dfinity.atlassian.net/browse/NNS1-3922)

# Changes

- New store for the viewport.

# Tests

- Unit tests.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-3922]:
https://dfinity.atlassian.net/browse/NNS1-3922?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ